Extract seed determiner
