+++ pkg/scheduler/controller/shoot/determiner_test.go
@@ -31,10 +31,10 @@ var _ = Describe("Scheduler_Control", func() {
 		ctrl             *gomock.Controller
 		fakeGardenClient client.Client
 
-		reconciler   *Reconciler
-		cloudProfile *gardencorev1beta1.CloudProfile
-		seed         *gardencorev1beta1.Seed
-		shoot        *gardencorev1beta1.Shoot
+		seedDeterminer *SeedDeterminer
+		cloudProfile   *gardencorev1beta1.CloudProfile
+		seed           *gardencorev1beta1.Seed
+		shoot          *gardencorev1beta1.Shoot
 
 		schedulerConfiguration config.SchedulerConfiguration
 
@@ -123,9 +123,9 @@ var _ = Describe("Scheduler_Control", func() {
 	})
 
 	JustBeforeEach(func() {
-		reconciler = &Reconciler{
-			Client: fakeGardenClient,
-			Config: schedulerConfiguration.Schedulers.Shoot,
+		seedDeterminer = &SeedDeterminer{
+			Client:   fakeGardenClient,
+			Strategy: schedulerConfiguration.Schedulers.Shoot.Strategy,
 		}
 	})
 
@@ -149,7 +149,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(bestSeed.Name).To(Equal(seed.Name))
 		})
@@ -169,7 +169,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, shoot)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, &secondShoot)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(bestSeed.Name).To(Equal(secondSeed.Name))
 		})
@@ -186,7 +186,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, &secondSeed)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, shoot)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(bestSeed.Name).To(Equal(secondSeed.Name))
 		})
@@ -199,7 +199,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).To(HaveOccurred())
 			Expect(bestSeed).To(BeNil())
 		})
@@ -217,7 +217,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, shoot)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).To(MatchError("none of the 1 seeds has at least 3 zones for hosting a shoot control plane with failure tolerance type 'zone'"))
 			Expect(bestSeed).To(BeNil())
 		})
@@ -237,7 +237,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, shoot)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).To(MatchError("none of the 1 seeds has at least 3 zones for hosting a shoot control plane with failure tolerance type 'zone'"))
 			Expect(bestSeed).To(BeNil())
 		})
@@ -253,7 +253,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, &multiZonalSeed)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, shoot)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(bestSeed.Name).To(Equal(multiZonalSeed.Name))
 		})
@@ -267,7 +267,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, &multiZonalSeed)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, shoot)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(bestSeed.Name).To(Equal(multiZonalSeed.Name))
 		})
@@ -296,7 +296,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(bestSeed).NotTo(BeNil())
 		})
@@ -310,7 +310,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(bestSeed).NotTo(BeNil())
 		})
@@ -324,7 +324,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(bestSeed).NotTo(BeNil())
 		})
@@ -344,7 +344,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(bestSeed).NotTo(BeNil())
 		})
@@ -355,7 +355,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).To(HaveOccurred())
 			Expect(bestSeed).To(BeNil())
 		})
@@ -372,7 +372,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).To(HaveOccurred())
 			Expect(bestSeed).To(BeNil())
 		})
@@ -390,7 +390,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).To(HaveOccurred())
 			Expect(bestSeed).To(BeNil())
 		})
@@ -411,7 +411,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(bestSeed.Name).To(Equal(seedName))
 		})
@@ -423,7 +423,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(bestSeed.Name).To(Equal(seedName))
 			// verify that shoot is in another region than the seed
@@ -451,7 +451,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, &secondSeed)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, &thirdSeed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(bestSeed.Name).To(Equal(secondSeed.Name))
 			// verify that shoot is in another region than the chosen seed
@@ -473,7 +473,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, shoot)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, &secondShoot)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(bestSeed.Name).To(Equal(secondSeed.Name))
 		})
@@ -513,7 +513,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, oldSeedEnvironment1)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, newSeedEnvironment2)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, testShoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, testShoot)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(bestSeed.Name).To(Equal(newSeedEnvironment2.Name))
 		})
@@ -563,7 +563,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, newSeedEnvironment2)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, newSeedEnvironment3)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, testShoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, testShoot)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(bestSeed.Name).To(Equal(newSeedEnvironment3.Name))
 		})
@@ -594,7 +594,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, &secondShoot)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, &thirdShoot)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(bestSeed.Name).To(Equal(secondSeed.Name))
 		})
@@ -615,7 +615,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(bestSeed.Name).To(Equal(seedName))
 		})
@@ -631,7 +631,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(bestSeed.Name).To(Equal(seedName))
 		})
@@ -650,7 +650,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, shoot)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, &secondShoot)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(bestSeed.Name).To(Equal(secondSeed.Name))
 		})
@@ -667,7 +667,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).To(HaveOccurred())
 			Expect(bestSeed).To(BeNil())
 		})
@@ -679,7 +679,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).To(HaveOccurred())
 			Expect(bestSeed).To(BeNil())
 		})
@@ -696,7 +696,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, shoot)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, &secondShoot)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).To(HaveOccurred())
 			Expect(bestSeed).To(BeNil())
 		})
@@ -708,7 +708,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).To(HaveOccurred())
 			Expect(bestSeed).To(BeNil())
 		})
@@ -719,7 +719,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).To(HaveOccurred())
 			Expect(bestSeed).To(BeNil())
 		})
@@ -735,7 +735,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).To(HaveOccurred())
 			Expect(bestSeed).To(BeNil())
 		})
@@ -752,7 +752,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).To(HaveOccurred())
 			Expect(bestSeed).To(BeNil())
 		})
@@ -762,7 +762,7 @@ var _ = Describe("Scheduler_Control", func() {
 
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).To(HaveOccurred())
 			Expect(bestSeed).To(BeNil())
 		})
@@ -778,7 +778,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).To(HaveOccurred())
 			Expect(bestSeed).To(BeNil())
 		})
@@ -789,7 +789,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).To(HaveOccurred())
 			Expect(bestSeed).To(BeNil())
 		})
@@ -804,7 +804,7 @@ var _ = Describe("Scheduler_Control", func() {
 			Expect(fakeGardenClient.Create(ctx, cloudProfile)).To(Succeed())
 			Expect(fakeGardenClient.Create(ctx, seed)).To(Succeed())
 
-			bestSeed, err := reconciler.determineSeed(ctx, log, shoot)
+			bestSeed, err := seedDeterminer.Determine(ctx, log, shoot)
 			Expect(err).To(HaveOccurred())
 			Expect(bestSeed).To(BeNil())
 		})
